requirements.txt
# Flask Application Dependencies for GCP Cloud Run
Flask==2.3.3
Werkzeug==2.3.7

# PyTorch and Computer Vision
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1

# Scientific Computing
numpy==1.24.3

# HTTP and JSON handling
requests==2.31.0

# Logging and utilities
python-json-logger==2.0.7

# Production WSGI server
gunicorn==21.2.0

# For handling base64 and image processing
opencv-python-headless==4.8.1.78

# Memory optimization for Cloud Run
psutil==5.9.6